ros2_parser/pyproject.toml
[project]
name = "ros2-indexer"
version = "0.2.0"
description = "Index ROS2 package metadata into JSON for MCP server consumption"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "typer>=0.9",
    "fastmcp>=3.0",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[project.scripts]
ros2-indexer = "ros2_indexer.cli:app"
ros2-index-mcp = "mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ros2_indexer", "src/mcp_server"]

[tool.pytest.ini_options]
testpaths = ["tests", "src/mcp_server/tests"]
markers = [
    "integration: tests that require network access",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"src/ros2_indexer/cli.py" = ["B008"]  # typer.Option() in defaults is standard Typer pattern

[tool.ruff.lint.isort]
known-first-party = ["ros2_indexer", "mcp_server"]
